+++ R/playlist.R
@@ -56,11 +56,11 @@ getPlaylist <- function(opts = antaresRead::simOptions()) {
   }
   
   # otherwise, update the vector of mc_years by removing disabled years
+  playlist_update_type <- names(parameters$playlist)
   playlist_update_value <- parameters$playlist
-  for (type in names(playlist_update_value)){
+  for (type in playlist_update_type){
     if (class(playlist_update_value[[type]]) %in% "list") playlist_update_value[[type]] <- unlist(playlist_update_value[[type]], recursive = F)
   }
-  playlist_update_type <- names(playlist_update_value)
   
   # untouched playlist - no modification have been made
   if (length(playlist_update_type) == 0) {